Renamed files to lowercase pattern
