+++ package.json
@@ -21,10 +21,13 @@
   "author": "",
   "license": "ISC",
   "devDependencies": {
-    "babel-cli": "^6.26.0",
+
     "babel-core": "^6.26.3",
+    "babel-loader": "^8.1.0",
     "babel-preset-env": "^1.7.0",
-    "babel-preset-es2015": "^6.24.1",
+    "webpack": "^4.43.0",
+    "webpack-dev-server": "^3.11.0",
+
     "cssmin": "^0.4.3",
     "htmlmin": "0.0.7",
     "node-sass": "^4.14.1",
@@ -33,8 +36,10 @@
     "rimraf": "^3.0.2",
     "uglifyjs": "2.4.11",
     "usemin-cli": "^0.6.0",
-    "webpack": "^4.43.0",
-    "webpack-cli": "^3.3.11"
+
+    "webpack-cli": "^3.3.11",
+    "babel-preset-es2015": "^6.24.1",
+    "babel-cli": "^6.26.0"
   },
   "dependencies": {
     "bootstrap": "^4.4.1",
