+++ scripts/common/sound-effects.js
